+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-ssr.js
@@ -1,4 +1,3 @@
-// const { shallow } = require(`enzyme`)
 const { onRenderBody } = require(`../gatsby-ssr`)
 
 let headComponents
@@ -36,7 +35,25 @@ describe(`gatsby-plugin-manifest`, () => {
   })
 
   it(`Adds "shortcut icon" and "manifest" links and "theme_color" meta tag to head`, () => {
-    onRenderBody(ssrArgs, { icon: true, theme_color: `#000000` })
+    onRenderBody(ssrArgs, {
+      icon: true,
+      theme_color: `#000000`,
+    })
+    expect(headComponents).toMatchSnapshot()
+  })
+
+  it(`Adds link favicon tag if "include_favicon" is set to true`, () => {
+    onRenderBody(ssrArgs, { icon: true, include_favicon: true })
+    expect(headComponents).toMatchSnapshot()
+  })
+
+  it(`Adds link favicon if "include_favicon" option is not provided`, () => {
+    onRenderBody(ssrArgs, { icon: true })
+    expect(headComponents).toMatchSnapshot()
+  })
+
+  it(`Does not add a link favicon if "include_favicon" option is set to false`, () => {
+    onRenderBody(ssrArgs, { icon: true, include_favicon: false })
     expect(headComponents).toMatchSnapshot()
   })
 
@@ -45,12 +62,11 @@ describe(`gatsby-plugin-manifest`, () => {
     expect(headComponents).toMatchSnapshot()
   })
 
-  describe(`Creates legacy apple touch links if opted in`, () => {
+  describe(`Creates legacy apple touch links`, () => {
     it(`Using default set of icons`, () => {
       onRenderBody(ssrArgs, {
         icon: true,
         theme_color: `#000000`,
-        legacy: true,
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -59,7 +75,38 @@ describe(`gatsby-plugin-manifest`, () => {
       onRenderBody(ssrArgs, {
         icon: true,
         theme_color: `#000000`,
-        legacy: true,
+        icons: [
+          {
+            src: `/favicons/android-chrome-48x48.png`,
+            sizes: `48x48`,
+            type: `image/png`,
+          },
+          {
+            src: `/favicons/android-chrome-512x512.png`,
+            sizes: `512x512`,
+            type: `image/png`,
+          },
+        ],
+      })
+      expect(headComponents).toMatchSnapshot()
+    })
+  })
+
+  describe(`Does not create legacy apple touch links`, () => {
+    it(`If "legacy" options is false and using default set of icons`, () => {
+      onRenderBody(ssrArgs, {
+        icon: true,
+        theme_color: `#000000`,
+        legacy: false,
+      })
+      expect(headComponents).toMatchSnapshot()
+    })
+
+    it(`If "legacy" options is false and using user specified list of icons`, () => {
+      onRenderBody(ssrArgs, {
+        icon: true,
+        theme_color: `#000000`,
+        legacy: false,
         icons: [
           {
             src: `/favicons/android-chrome-48x48.png`,
@@ -83,7 +130,6 @@ describe(`gatsby-plugin-manifest`, () => {
     onRenderBody(ssrArgs, {
       icon: true,
       theme_color: `#000000`,
-      legacy: true,
     })
 
     headComponents