+++ packages/gatsby-plugin-postcss/src/gatsby-node.js
@@ -67,7 +67,7 @@ exports.onCreateWebpackConfig = (
 
   if (!isSSR) {
     postcssRule.use.unshift(loaders.miniCssExtract())
-    postcssRuleModules.use.unshift(loaders.miniCssExtract())
+    postcssRuleModules.use.unshift(loaders.miniCssExtract({ hmr: false }))
   }
 
   const postcssRules = { oneOf: [] }