+++ starters/blog/src/components/Bio.js
@@ -1,8 +1,8 @@
-import React from 'react'
-import { StaticQuery, graphql } from 'gatsby'
-import Image from 'gatsby-image'
+import React from "react"
+import { StaticQuery, graphql } from "gatsby"
+import Image from "gatsby-image"
 
-import { rhythm } from '../utils/typography'
+import { rhythm } from "../utils/typography"
 
 function Bio() {
   return (
@@ -25,6 +25,9 @@ function Bio() {
                 marginBottom: 0,
                 minWidth: 50,
                 borderRadius: `100%`,
+              }}
+              imgStyle={{
+                borderRadius: `50%`,
               }}
             />
             <p>