+++ packages/gatsby-plugin-react-css-modules/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-css-modules/compare/gatsby-plugin-react-css-modules@2.0.5...gatsby-plugin-react-css-modules@2.0.6) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-react-css-modules
+
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-css-modules/compare/gatsby-plugin-react-css-modules@2.0.4...gatsby-plugin-react-css-modules@2.0.5) (2019-01-23)
+
+**Note:** Version bump only for package gatsby-plugin-react-css-modules
+
 <a name="2.0.4"></a>
 
 ## [2.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-css-modules/compare/gatsby-plugin-react-css-modules@2.0.3...gatsby-plugin-react-css-modules@2.0.4) (2018-11-29)